# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
    title: OrigCMS API
    description: OrigCMS is a modern Content Management System for video platforms.
    contact:
        name: GodCong
        url: https://github.com/origadmin
        email: waitforadding@gmail.com
    license:
        name: MIT
        url: https://origadmin/application/origcms/blob/master/LICENSE
    version: Version from annotation
servers:
    - url: http://localhost:18080
    - url: https://localhost:18080
paths:
...
    /api/v1/tags:
        get:
            tags:
                - TagService
            operationId: TagService_ListTags
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: page_size
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: page_token
                  in: query
                  schema:
                    type: string
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.ListTagsResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        post:
            tags:
                - TagService
            operationId: TagService_CreateTag
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.media.CreateTagRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.CreateTagResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /api/v1/tags/{id}:
        get:
            tags:
                - TagService
            operationId: TagService_GetTag
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.GetTagResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        delete:
            tags:
                - TagService
            operationId: TagService_DeleteTag
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.DeleteTagResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /api/v1/tags/{tag.id}:
        put:
            tags:
                - TagService
            operationId: TagService_UpdateTag
            parameters:
                - name: tag.id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.media.UpdateTagRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.UpdateTagResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
components:
    schemas:
...
        api.v1.services.media.CreateTagRequest:
            type: object
            properties:
                tag:
                    $ref: '#/components/schemas/api.v1.services.types.Tag'
        api.v1.services.media.CreateTagResponse:
            type: object
            properties:
                tag:
                    $ref: '#/components/schemas/api.v1.services.types.Tag'
...
        api.v1.services.media.DeleteTagResponse:
            type: object
            properties: {}
...
        api.v1.services.media.GetTagResponse:
            type: object
            properties:
                tag:
                    $ref: '#/components/schemas/api.v1.services.types.Tag'
...
        api.v1.services.media.ListTagsResponse:
            type: object
            properties:
                total:
                    type: integer
                    format: int32
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.v1.services.types.Tag'
                page:
                    type: integer
                    format: int32
                page_size:
                    type: integer
                    format: int32
                next_page_token:
                    type: string
                total_pages:
                    type: integer
                    format: int32
...
        api.v1.services.media.UpdateTagRequest:
            type: object
            properties:
                tag:
                    $ref: '#/components/schemas/api.v1.services.types.Tag'
                update_mask:
                    type: string
                    format: field-mask
        api.v1.services.media.UpdateTagResponse:
            type: object
            properties:
                tag:
                    $ref: '#/components/schemas/api.v1.services.types.Tag'
...
        api.v1.services.types.Tag:
            type: object
            properties:
                id:
                    type: string
                create_time:
                    type: string
                    format: date-time
                update_time:
                    type: string
                    format: date-time
                name:
                    type: string
                slug:
                    type: string
                description:
                    type: string
                color:
                    type: string
                status:
                    type: integer
                    format: enum
                media_count:
                    type: string
                listings_thumbnail:
                    type: string
            description: Tag is the model entity for Tag.
...
        google.protobuf.Any:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        google.rpc.Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/google.protobuf.Any'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
    securitySchemes:
        Basic:
            type: http
            scheme: basic
        Bearer:
            type: apiKey
            name: Authorization
            in: header
...